Consolidate checks and update documentation for #18
